Copy known_hosts only when it doesn't already exist

+++ .devcontainer/local-features/ssh/bin/postStartScript.sh
@@ -4,6 +4,11 @@
 sudo chmod 666 /run/host-services/ssh-auth.sock
 
 ## Copy known_hosts to reduce SSH prompts
-mkdir -p ~/.ssh
-cp ~/.ssh-host/known_hosts ~/.ssh/known_hosts
-chmod 600 ~/.ssh/known_hosts
+if [ ! -e ~/.ssh/known_hosts ]
+then
+  echo "Copying known_hosts from ~/.ssh-host to ~/.ssh"
+
+  mkdir -p ~/.ssh
+  cp ~/.ssh-host/known_hosts ~/.ssh/known_hosts
+  chmod 600 ~/.ssh/known_hosts
+fi
